[AMB-4443] bugFix fetch bond balance
