+++ src/features/kosmos/context/markets/markets.context.ts
@@ -2,8 +2,13 @@ import { useCallback, useState } from 'react';
 import { ethers } from 'ethers';
 import { createContextSelector } from '@utils/createContextSelector';
 import { Token, TxType } from '@features/kosmos/types';
+import { getBalanceOf } from '@features/kosmos/lib/contracts';
+import { NumberUtils } from '@utils/number';
+import { useWallet } from '@hooks';
 
 const KosmosMarketsContext = () => {
+  const { wallet } = useWallet();
+
   const [tokens, setTokens] = useState<Token[]>([]);
   const [isTokensLoading, setIsTokensLoading] = useState(false);
   const [isMarketTooltipVisible, setIsMarketTooltipVisible] = useState(false);
@@ -38,6 +43,29 @@ const KosmosMarketsContext = () => {
     setBnBalance(null);
   }, []);
 
+  const [bondBalance, setBondBalance] = useState('');
+  const fetchBondBalance = async (qouteToken: Token | undefined) => {
+    try {
+      setIsBalanceFetching(true);
+      const bnBalance = await getBalanceOf({
+        ownerAddress: wallet?.address ?? '',
+        token: qouteToken
+      });
+
+      const formattedBalance = NumberUtils.limitDecimalCount(
+        ethers.utils.formatEther(bnBalance?._hex),
+        2
+      );
+
+      if (formattedBalance !== bondBalance) {
+        setBondBalance(formattedBalance);
+        setBnBalance(bnBalance);
+      }
+    } finally {
+      setIsBalanceFetching(false);
+    }
+  };
+
   return {
     tokens,
     setTokens,
@@ -57,7 +85,9 @@ const KosmosMarketsContext = () => {
     setTransactions,
     claimedOrderIds,
     onAppendClaimedOrderId,
-    reset
+    reset,
+    bondBalance,
+    fetchBondBalance
   };
 };
